+++ NEWS.md
@@ -1,5 +1,10 @@
 # Changes will be documented here
 
+## Version 0.1.4
+
+- Message returned if supplies a numeric `location_var` with fewer than 5 
+unique values. This warns of possible problems calculating Moran's I.
+
 ## Version 0.1.3
 
 - `tlsl` argument added to `monadic_spatial_weights`. If `TRUE` then